+++ src/test/java/com/example/jpa/nPlusOne/FetchJoinTest.java
@@ -1,6 +1,6 @@
 package com.example.jpa.nPlusOne;
 
-import com.example.jpa.nPlusOne.service.SolutionService;
+import com.example.jpa.nPlusOne.service.FetchJoinService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,32 +9,32 @@ import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 @Transactional
-public class SolutionTest {
+public class FetchJoinTest {
 
     @Autowired
-    private SolutionService solutionService;
+    private FetchJoinService fetchJoinService;
 
     @DisplayName("fetch join 기반 N+1 문제 해결 테스트")
     @Test
     void testFetchJoin() {
-        solutionService.findAllMembersByFetchJoin();
+        fetchJoinService.findAllMembersByFetchJoin();
     }
 
     @DisplayName("JDBC 기반 join 문법 직접 활용을 통한 N+1 문제 해결 테스트")
     @Test
     void testJdbc() {
-        solutionService.findAllMembersByJdbc();
+        fetchJoinService.findAllMembersByJdbc();
     }
 
     @DisplayName("페이징과 fetch join을 같이 사용하면 메모리 아웃 예외가 발생할 수 있다")
     @Test
     void testFetchJoinWithPaging() {
-        solutionService.findAllMembersWithPaging();
+        fetchJoinService.findAllMembersWithPaging();
     }
 
     @DisplayName("EntityGraph 기반 N+1 문제 해결 테스트")
     @Test
     void testEntityGraph() {
-        solutionService.findAllMembersWithEntityGraph();
+        fetchJoinService.findAllMembersWithEntityGraph();
     }
 }
